Update DependentHandle on Mono runtime
